添加选择完微信号才出现相应内容

+++ backend/base/BackendController.php
@@ -89,19 +89,22 @@ class BackendController extends Controller {
 	public function beforeAction($action) {
 		parent::beforeAction ( $action );
 		// 访问非菜单里的action时，菜单保持打开(添加角色时角色管理保持打开状态)
-		$cache = Yii::$app->getCache ();
-		$cache->delete('accounts');
-		$results = [];
-		if (empty ( $cache->get ( 'accounts' ) )) {
-			$user = yii::$app->user->getIdentity();
-			$accounts = PublicAccount::find()
-			->joinWith('weixinUsers')->where(['user_id' => $user->id])->all();
-			foreach($accounts as $account){
-				$results[$account->type][] = $account;
-			}
-			Yii::$app->getCache()->set('accounts', $results);
-		}
 		
+		$user = yii::$app->user->getIdentity();
+		if(!empty($user)){
+			$cache = Yii::$app->getCache ();
+			$cache->delete('accounts');
+			$results = [];
+			if (empty ( $cache->get ( 'accounts' ) )) {
+					
+				$accounts = PublicAccount::find()
+				->joinWith('weixinUsers')->where(['user_id' => $user->id])->all();
+				foreach($accounts as $account){
+					$results[$account->type][] = $account;
+				}
+				Yii::$app->getCache()->set('accounts', $results);
+			}
+		}
 		$refferroute = Yii::$app->request->referrer;
 		$_referrer = parse_url ( $refferroute );
 		Yii::$app->session->set ( 'referrerroute', $_referrer ['path'] );
